pyproject.toml
[tool.poetry]
name = "cs50-final-project"
version = "0.1.0"
description = "A carbon footprint calculator designed to make it easy to make better decisions for the planet."
authors = ["Bianca Silva"]
license = "Open Sourced"
readme = "README.md"
packages = [{include = "cs50_final_project"}]

[tool.poetry.dependencies]
python = "3.10.6"
Flask = "^2.2.2"
Flask-Session = "^0.4.0"
cs50 = "^9.2.2"
gunicorn = "^20.1.0"
requests = "^2.28.1"
psycopg2 = "^2.9.3"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
